주석 추가 및 HealthStatus 보완

+++ src/main/java/com/dmc/bootcamp/dto/response/UserResponse.java
@@ -13,8 +13,8 @@ public class UserResponse {
     private Character gender;
     private LocalDate birthday;
     private String diseaseInfo;
-    private float height;
-    private float weight;
+    private Double height;
+    private Double weight;
     private  String address;
 
     public UserResponse(AppUser user){
